LDSNetApi/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.6.0
asttokens==2.4.1
astunparse==1.6.3
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.0
debugpy==1.8.0
decorator==4.4.2
executing==2.0.1
fastapi==0.115.0
flatbuffers==24.3.25
gast==0.6.0
google-pasta==0.2.0
grpcio==1.66.1
h11==0.14.0
h5py==3.11.0
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
immutabledict==4.2.0
ipykernel==6.28.0
ipython==8.19.0
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.6.1
keras==3.5.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.4.1
moviepy==1.0.3
namex==0.0.8
nest-asyncio==1.5.8
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
ortools==9.9.3963
packaging==24.1
pandas==2.2.3
parso==0.8.3
pillow==10.4.0
platformdirs==4.1.0
proglog==0.1.10
prompt-toolkit==3.0.43
protobuf==4.25.5
psutil==5.9.7
pure-eval==0.2.2
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-multipart==0.0.10
pytz==2024.2
pyzmq==25.1.2
requests==2.32.3
rich==13.8.1
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.38.6
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tornado==6.4
tqdm==4.66.5
traitlets==5.14.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.30.6
wcwidth==0.2.12
Werkzeug==3.0.4
wrapt==1.16.0
